+++ .lintstagedrc.js
@@ -14,7 +14,7 @@ const Backtrack = require('@backtrack/core');
 const { configManager } = new Backtrack();
 
 const lintStaged = {
-    '*.{js,mjs,jsx,ts,tsx,json,scss,less,css,md,yml,yaml}': [
+    '*.{js,jsx,ts,tsx,mjs,json,scss,less,css,md,yml,yaml}': [
         'prettier --write',
         'git add',
     ],