+++ src/ex-config.js
@@ -366,5 +366,4 @@ class ExConfig {
     }
 }
 
-// eslint-disable-next-line import/no-default-export
 export default ExConfig;